% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_selector.R
\name{beta_selector}
\alias{beta_selector}
\title{Beta Selector}
\usage{
beta_selector(t, alpha = 0.05, beta_grid = seq(0.1, 1, 0.05))
}
\arguments{
\item{t}{A numeric vector of angles.}

\item{alpha}{Level for pettitt.test.}

\item{beta_grid:}{A numeric vector with values between 0 y 1. It represents a grid of beta values.}
}
\value{
A list width the beta selected, the p.value and the vector of standard deviations.
}
\description{
Selects \eqn{\beta} in a data driven way for a circular sample.
}
\examples{

library(circular)
tita = as.numeric(rmixvonmises(1000, mu1=pi, mu2=0, kappa1=10, kappa2=10, prop=0.3))
beta_selector(t=tita)

}
