% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld.R
\name{cld}
\alias{cld}
\title{Circular Local Depth}
\usage{
cld(t, beta)
}
\arguments{
\item{t}{A numeric vector of angles to calculate the depth.}

\item{beta}{Locality parameter.}
}
\value{
a numeric vector with the circular depth for each value of t.
}
\description{
Calculates the circular depth for a sample.
}
\examples{

library(circular)
tita = as.numeric(rvonmises(100, mu=pi, kappa=5))
cld(t=tita, beta=0.2)
}
