% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_v3.R
\name{icd_v3}
\alias{icd_v3}
\title{Integrated Circular Depth (version 3)}
\usage{
icd_v3(traj, beta = 0.2, probs = seq(0, 1, 0.1), polar = FALSE)
}
\arguments{
\item{traj}{A list which contains the trajectories.}

\item{beta}{Locality level.}

\item{probs}{Probabilities for the grid construction.}

\item{polar}{If TRUE is assumed that trajectories are in polar coordinates. Contrary, is assumed long/lat coordinates.}
}
\value{
a numeric vector which contains the depth for each trajectory.
}
\description{
Integrated Circular Depth (version 3)
}
