% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_v5.R
\name{icd_v5}
\alias{icd_v5}
\title{Integrated Circular Depth (version 5 with weights)}
\usage{
icd_v5(
  traj,
  beta = "automatic",
  probs = seq(0, 1, 0.1),
  type = "geographical",
  weight = TRUE
)
}
\arguments{
\item{traj}{A list which contains the trajectories.}

\item{beta}{Locality level numeric between 0 and 1. Could be set as "automatic".}

\item{probs}{Sequense of probabilities for grid construction.}

\item{type}{The trajectory type of coordinates geographical or cartesian. If "geographical" made correction for lat and long coordinates.}

\item{weight}{If TRUE a weight in each radius is applied.}
}
\value{
a numeric vector which contains the depth for each trajectory.
}
\description{
Integrated Circular Depth (version 5 with weights)
}
