% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld2.R
\name{lambda_plus}
\alias{lambda_plus}
\title{Circular Local Depth}
\usage{
lambda_plus(pos, lambda, n)
}
\arguments{
\item{x}{A numeric vector of angles to calculate the depth.}

\item{data}{A numeric vector that will be used to calculate depth.}

\item{beta}{Locality parameter.}
}
\value{
A numeric vector with the circular depth for each value of x.
}
\description{
Calculates the circular local depth for a grid of values based on a sample.
}
\examples{

library(circular)
tita = as.numeric(rvonmises(100, mu=pi, kappa=5))
cld2(x=0, data=tita, beta=0.2)
}
